workflows: Use new test-repository features
Use three new features of test-repository:

* update_base_url

During publish workflow we want to test that a client can update from
current GCS (tuf-repo.cdn.sigstage.dev) to Pages
(sigstore.github.io/root-signing-staging/).

This test is done during cron runs as well but then it's usually
not relevant (as the repos contents typically match).

* valid_days

We want a failure if either Pages or GCS repository is invalid in 3
days.

Note: sigstore-probers alerts when repository is invalid in 2 days

* offline_valid_days

We want a failure if root or targets in either Pages or GCS repository
expires within 30 days.

Note: sigstore-probers alerts when offline roles expire in 15 days

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>