Avoid running signing-event workflow in forks
This came up in root-signing prod signing event: Signers may have
actions enabled in their forks. This leads to signing event running in
their fork. The results may be very strange as they depend on the state
of main (which on forks may be years old).

Avoid running the signing-event workflow in forks.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>