tests: Lower the required validity time
currently the signing periods are 32 days.
If we require metadata to to be valid for 30 days, then there is only
2 days signing time before there is a test failure.

Lower this to 16 days: That way we get a test failure if the signing has
not happened in half the signing period -- this is good in case of e.g.
signing event creation failing.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>