workflows: Fix issue with future validity check
This is actually never true:
    github.event_name == 'workflow_call'
as the event name in workflow_called workflows is the calling
workflows event name.

What we want is to distinguish publishing case from other cases
(cron and manual dispatch): in this case we do not want "future"
validity check.

Using workflow name for this is a bit of a hack but at least it works
unlike the current if-clause.

Fixes #216

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>